lsp: remove on_setup_handler
This aligns a bit better with the new changes in neovim LSP ecosystem.
nvim-lspconfig is upstreaming most of its "framework" bits to
vim.lsp.config. As part of this, most of the complexity of the existing
configuration is gone.

Eventually we need to move from lspconfig's hooks to vim.lsp hooks (are
there any?) so this commit is just a start to make LSP working again...